refactor: extract CLI package with git-friendly moves
Move generic CLI utilities to @mongodb-js/mcp-cli package:
- git mv transports/constants.ts -> packages/cli/src/transports/
- git mv common/config/configUtils.ts -> packages/cli/src/config/
- Create CLI package scaffold (package.json, tsconfig.json, index.ts)
- Update mongodb-mcp-server to depend on @mongodb-js/mcp-cli
- Update all imports to use new package
- Create wrapper configUtils.ts in mongodb-mcp-server for MongoDB-specific helpers
