+++ packages/mongodb-mcp-server/src/common/config/configUtils.ts
@@ -1,43 +1,18 @@
 import path from "path";
 import os from "os";
 
-/// Custom logic function to apply the override value.
-/// Returns the value to use (which may be transformed from newValue).
-/// Should throw an error if the override cannot be applied.
-export type CustomOverrideLogic = (oldValue: unknown, newValue: unknown) => unknown;
-
-/**
- * Defines how a config field can be overridden via HTTP headers or query parameters.
- */
-export type OverrideBehavior =
-    /// Cannot be overridden via request
-    | "not-allowed"
-    /// Can be completely replaced
-    | "override"
-    /// Values are merged (for arrays)
-    | "merge"
-    | CustomOverrideLogic;
-
-/**
- * Metadata for config schema fields.
- */
-export type ConfigFieldMeta = {
-    /**
-     * Custom description for the default value, used when generating documentation.
-     */
-    defaultValueDescription?: string;
-    /**
-     * Marks the field as containing sensitive/secret information, used for MCP Registry.
-     * Secret fields will be marked as secret in environment variable definitions.
-     */
-    isSecret?: boolean;
-    /**
-     * Defines how this config field can be overridden via HTTP headers or query parameters.
-     * Defaults to "not-allowed" for security.
-     */
-    overrideBehavior?: OverrideBehavior;
-    [key: string]: unknown;
-};
+// Re-export generic utilities from mcp-cli
+export {
+    commaSeparatedToArray,
+    parseBoolean,
+    oneWayOverride,
+    onlyLowerThanBaseValueOverride,
+    onlyStricterLogLevelOverride,
+    onlySubsetOfBaseValueOverride,
+    type CustomOverrideLogic,
+    type OverrideBehavior,
+    type ConfigFieldMeta,
+} from "@mongodb-js/mcp-cli";
 
 export function getLocalDataPath(): string {
     return process.platform === "win32"
@@ -52,122 +27,4 @@ export function getLogPath(): string {
 
 export function getExportsPath(): string {
     return path.join(getLocalDataPath(), "mongodb-mcp", "exports");
-}
-
-export function commaSeparatedToArray<T extends string[]>(str: string | string[] | undefined): T | undefined {
-    if (str === undefined) {
-        return undefined;
-    }
-
-    if (typeof str === "string") {
-        return str
-            .split(",")
-            .map((e) => e.trim())
-            .filter((e) => e.length > 0) as T;
-    }
-
-    if (str.length === 1) {
-        return str[0]
-            ?.split(",")
-            .map((e) => e.trim())
-            .filter((e) => e.length > 0) as T;
-    }
-
-    return str as T;
-}
-
-/**
- * Preprocessor for boolean values that handles string "false"/"0" correctly.
- * Zod's coerce.boolean() treats any non-empty string as true, which is not what we want.
- */
-export function parseBoolean(val: unknown): unknown {
-    if (val === undefined) {
-        return undefined;
-    }
-    if (typeof val === "string") {
-        if (val === "false") {
-            return false;
-        }
-        if (val === "true") {
-            return true;
-        }
-        throw new Error(`Invalid boolean value: ${val}`);
-    }
-    if (typeof val === "boolean") {
-        return val;
-    }
-    if (typeof val === "number") {
-        return val !== 0;
-    }
-    return !!val;
-}
-
-/** Allow overriding only to the allowed value */
-export function oneWayOverride<T>(allowedValue: T): CustomOverrideLogic {
-    return (oldValue, newValue) => {
-        // Only allow override if setting to allowed value or current value
-        if (newValue === oldValue) {
-            return newValue;
-        }
-        if (newValue === allowedValue) {
-            return newValue;
-        }
-        throw new Error(`Can only set to ${String(allowedValue)}`);
-    };
-}
-
-/** Allow overriding only to a value lower than the specified value */
-export function onlyLowerThanBaseValueOverride(): CustomOverrideLogic {
-    return (oldValue, newValue) => {
-        if (typeof oldValue !== "number") {
-            throw new Error(`Unsupported type for base value for override: ${typeof oldValue}`);
-        }
-        if (typeof newValue !== "number") {
-            throw new Error(`Unsupported type for new value for override: ${typeof newValue}`);
-        }
-        if (newValue >= oldValue) {
-            throw new Error(`Can only set to a value lower than the base value`);
-        }
-        return newValue;
-    };
-}
-
-/**
- * Allow overriding a log level only to a stricter (higher severity) value.
- * The ordered list must go from least to most severe.
- */
-export function onlyStricterLogLevelOverride(orderedLevels: readonly string[]): CustomOverrideLogic {
-    return (oldValue, newValue) => {
-        if (typeof oldValue !== "string" || typeof newValue !== "string") {
-            throw new Error(`Expected string log level values`);
-        }
-        const oldIdx = orderedLevels.indexOf(oldValue);
-        const newIdx = orderedLevels.indexOf(newValue);
-        if (oldIdx === -1 || newIdx === -1) {
-            throw new Error(`Unknown log level`);
-        }
-        if (newIdx < oldIdx) {
-            throw new Error(`Can only override to a stricter (higher severity) log level`);
-        }
-        return newValue;
-    };
-}
-
-/** Allow overriding only to a subset of an array but not a superset */
-export function onlySubsetOfBaseValueOverride(): CustomOverrideLogic {
-    return (oldValue, newValue) => {
-        if (!Array.isArray(oldValue)) {
-            throw new Error(`Unsupported type for base value for override: ${typeof oldValue}`);
-        }
-        if (!Array.isArray(newValue)) {
-            throw new Error(`Unsupported type for new value for override: ${typeof newValue}`);
-        }
-        if (newValue.length > oldValue.length) {
-            throw new Error(`Can only override to a subset of the base value`);
-        }
-        if (!newValue.every((value) => oldValue.includes(value))) {
-            throw new Error(`Can only override to a subset of the base value`);
-        }
-        return newValue as unknown;
-    };
 }
