+++ packages/mongodb-mcp-server/src/common/config/userConfig.ts
@@ -13,7 +13,7 @@ import {
 import { MCP_LOG_LEVELS } from "@mongodb-js/mcp-core";
 import { monitoringServerFeatureValues, previewFeatureValues } from "@mongodb-js/mcp-tools-mongodb";
 import { argMetadata, CliOptionsSchema as MongoshCliOptionsSchema } from "@mongosh/arg-parser/arg-parser";
-import { TRANSPORT_PAYLOAD_LIMITS } from "../../transports/constants.js";
+import { TRANSPORT_PAYLOAD_LIMITS } from "@mongodb-js/mcp-cli";
 
 export const configRegistry = z.registry<ConfigFieldMeta>();
